lingering renaming of Backend to Solver and removal of unnecessary Solver handler
