+++ chirho/dynamical/handlers/dynamical.py
@@ -16,15 +16,19 @@ from chirho.dynamical.internals.interruption import (
     concatenate,
     simulate_to_interruption,
 )
-from chirho.dynamical.ops import Trajectory
+from chirho.dynamical.ops.dynamical import Trajectory
 
 S = TypeVar("S")
 T = TypeVar("T")
 
 
+class Solver(pyro.poutine.messenger.Messenger):
+    def _pyro_simulate(self, msg) -> None:
+        # Overwrite the solver in the message with the enclosing solver when used as a context manager.
+        msg["kwargs"]["solver"] = self
+
+
 class SimulatorEventLoop(Generic[T], pyro.poutine.messenger.Messenger):
-    def __enter__(self):
-        return super().__enter__()
 
     # noinspection PyMethodMayBeStatic
     def _pyro_simulate(self, msg) -> None:
