+++ chirho/dynamical/handlers/solver.py
@@ -2,14 +2,8 @@ from __future__ import annotations
 
 import pyro
 
-from chirho.dynamical.ops import Solver
 
-
-class SolverHandler(pyro.poutine.messenger.Messenger):
-    def __init__(self, solver: Solver):
-        self.solver = solver
-        super().__init__()
-
+class Solver(pyro.poutine.messenger.Messenger):
     def _pyro_simulate(self, msg) -> None:
-        # Overwrite the solver in the message with the one we're handling.
-        msg["kwargs"]["solver"] = self.solver
+        # Overwrite the solver in the message with the enclosing solver when used as a context manager.
+        msg["kwargs"]["solver"] = self
