+++ tests/dynamical/test_handler_composition.py
@@ -11,8 +11,8 @@ from chirho.dynamical.handlers import (
     PointIntervention,
     SimulatorEventLoop,
 )
+from chirho.dynamical.handlers.ODE.solvers import TorchDiffEq
 from chirho.dynamical.ops import State, simulate
-from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 from chirho.observational.handlers.soft_conditioning import AutoSoftConditioning
 from tests.dynamical.dynamical_fixtures import (
     UnifiedFixtureDynamics,
