+++ tests/dynamical/test_static_observation.py
@@ -11,8 +11,8 @@ from chirho.dynamical.handlers import (
     PointObservation,
     SimulatorEventLoop,
 )
+from chirho.dynamical.handlers.ODE.solvers import TorchDiffEq
 from chirho.dynamical.ops import State, simulate
-from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 
 from .dynamical_fixtures import (
     UnifiedFixtureDynamics,
